Figure 4
Figure 4 displays the probability the problem is answered correctly across varying levels of difficulty, for each of the three problem types, separated by political ideology, and problem difficulty. Although the 9 panels are a bit complicated to take in, the main interactions tested in this study (3-ways between numeracy, political ideology, and problem type, and the 4-way interaction) are not seen in this figure. In fact, only the main effect of numeracy is clearly present (all interactions involving numeracy were not significant).